Fixed minor issues

+++ database/migrations/2022_04_25_211125_create_customer_trips_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->unsignedBigInteger('customer_id');
             $table->unsignedBigInteger('trip_id');
             $table->boolean('checkout');
-            $table->integer('rate')->nullable();
+            $table->decimal('rate')->nullable();
             $table->softDeletes();
             $table->timestamps();
         });
